+++ src/main/java/com/coupon/issuecouponservice/security/oauth/GoogleUserInfo.java
@@ -1,4 +1,4 @@
-package com.coupon.issuecouponservice.security;
+package com.coupon.issuecouponservice.security.oauth;
 
 import java.util.Map;
