+++ src/main/java/com/coupon/issuecouponservice/security/oauth/PrincipalOauth2UserService.java
@@ -3,10 +3,6 @@ package com.coupon.issuecouponservice.security.oauth;
 import com.coupon.issuecouponservice.domain.user.Role;
 import com.coupon.issuecouponservice.domain.user.User;
 import com.coupon.issuecouponservice.repository.user.UserRepository;
-import com.coupon.issuecouponservice.security.GoogleUserInfo;
-import com.coupon.issuecouponservice.security.KakaoUserInfo;
-import com.coupon.issuecouponservice.security.NaverUserInfo;
-import com.coupon.issuecouponservice.security.OAuth2UserInfo;
 import com.coupon.issuecouponservice.security.userdetails.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
